create error page test on delete student

+++ Project/src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -35,6 +35,9 @@ public class GradebookController {
 
     @GetMapping("/delete/student/{id}")
     public String deleteStudent(@PathVariable int id, Model m) {
+        if(!studentService.checkIfStudentIsNull(id)){
+            return "error";
+        }
         studentService.deleteStudent(id);
         Iterable<CollegeStudent> collegeStudents = studentService.getGradebook();
         m.addAttribute("students", collegeStudents);
